+++ lib/elastic/dsl/errors/base_error.rb
@@ -1,16 +1,11 @@
 module Elastic
   module DSL
     module Errors
-      class BaseError < Exception; end
+      class BaseError < StandardError; end
+      class InvalidArgument < BaseError; end
 
-      class NodeNotFound < Elastic::DSL::Errors::BaseError
-        attr_accessor :index
-
-        def intialize(index)
-          @index = index
-        end
-      end
-
+      class InvalidQuery < Elastic::DSL::Errors::BaseError; end
+      class NodeNotFound < Elastic::DSL::Errors::BaseError; end
     end
   end
 end