+++ spec/elastic/dsl/builders/utils_spec.rb
@@ -2,7 +2,7 @@ require 'spec_helper'
 
 describe Elastic::DSL::Builders::Utils do
 
-  let(:builder) { Class.new { extend Elastic::DSL::Builders::Utils } }
+  let(:utils) { Class.new { extend Elastic::DSL::Builders::Utils } }
 
   describe '.deeper_merge!' do
 
@@ -52,100 +52,112 @@ describe Elastic::DSL::Builders::Utils do
 
     describe 'deeper_merge' do
       it 'Merges two querys into a new hash' do
-        expect(merged = builder.deeper_merge(tree, override, :bool)).to eq(result)
+        expect(merged = utils.deeper_merge(tree, override, :bool)).to eq(result)
         expect(merged.object_id).not_to eq(result.object_id)
       end
     end
 
     describe 'deeper_merge!' do
       it 'Merges two querys' do
-        expect(builder.deeper_merge!(tree, override, :bool)).to eq(result)
+        expect(utils.deeper_merge!(tree, override, :bool)).to eq(result)
       end
     end
   end
 
   describe '.find_node' do
-    let(:root_node) { {query: {test: 'test'}} }
+
+    let(:node_value) {'value'}
+    let(:wrong_value) {'bad_value'}
+
+    let(:root_node) { {query: {
+      test: node_value,
+      veronica: nil,
+      nested_array: [{jimmy: wrong_value}, {mark: node_value}],
+      bad_array: [{jimmy: wrong_value}, wrong_value, {mark: node_value}],
+      nested_hash: {ashleigh: true},
+    }} }
 
     it 'raises ArgumentError with blank node_list' do
-      expect{builder.find_node([], root_node)}.to raise_error(ArgumentError)
+      expect{utils.find_node([], root_node)}.to raise_error(Elastic::DSL::Errors::InvalidArgument)
     end
 
-    it 'raises NodeNotFound error when it doesn\'t exist' do
-      expect{builder.find_node([:bob], root_node)}.to raise_error(Elastic::DSL::Errors::NodeNotFound)
+    it 'raises InvalidQuery if an array contains a non-hash' do
+      expect{utils.find_node([:query, :bad_array, :wrong_value], root_node)}.to raise_error(Elastic::DSL::Errors::InvalidQuery)
     end
 
-    it 'returns the node from an array of keys' do
-      expect(builder.find_node([:query], root_node)).to eq(root_node[:query])
+    context 'raises NodeNotFound Errors' do
+
+      it 'when a node doesn\'t exist in a hash' do
+        expect{utils.find_node([:query, :nested_hash, :not_a_node], root_node)}.to raise_error(Elastic::DSL::Errors::NodeNotFound)
+      end
+
+      it 'when a node doesn\'t exist in an array' do
+        expect{utils.find_node([:query, :nested_array, :not_a_node], root_node)}.to raise_error(Elastic::DSL::Errors::NodeNotFound)
+      end
+
+      it 'when the path ends prematurely' do
+        expect{utils.find_node([:query, :not_a_node, :jimmy], root_node)}.to raise_error(Elastic::DSL::Errors::NodeNotFound)
+      end
+
     end
 
-    it 'returns the node from a single key' do
-      expect(builder.find_node(:query, root_node)).to eq(root_node[:query])
+    context 'returns the node\'s value' do
+
+      it 'when queried with an array of keys' do
+        expect(utils.find_node([:query], root_node)).to eq(root_node[:query])
+      end
+
+      it 'when queried with a single key' do
+        expect(utils.find_node(:query, root_node)).to eq(root_node[:query])
+      end
+
+      it 'when it maps to nil' do
+        expect(utils.find_node([:query, :veronica], root_node)).to eq(nil)
+      end
+
+      it 'from a nested hash' do
+        expect(utils.find_node([:query, :test], root_node)).to eq(node_value)
+      end
+
+      it 'from a nested array' do
+        expect(utils.find_node([:query, :nested_array], root_node)).to eq(root_node[:query][:nested_array])
+      end
+
+      it 'from a nested hash in an array' do
+        expect(utils.find_node([:query, :nested_array, :mark], root_node)).to eq(node_value)
+      end
+
     end
 
-    it 'returns the nested node' do
-      expect(builder.find_node([:query, :test], root_node)).to eq(root_node[:query][:test])
-    end
   end
-
-  # Indifferent Access Hash
-  # describe '.find_node' do
-  #   let(:root_node) { {query: {test: 'test'}} }
-  #   # let(:root_node2) { HashWithIndifferentAccess.new(query: {test: 'test'}) }
-
-  #   it 'raises ArgumentError with blank node_list' do
-  #     expect{builder.find_node([], root_node)}.to raise_error(ArgumentError)
-  #   end
-
-  #   it 'raises NodeNotFound error when it doesn\'t exist' do
-  #     expect{builder.find_node(['bob'], root_node)}.to raise_error(Elastic::DSL::Errors::NodeNotFound)
-  #   end
-
-  #   it 'returns the node from an array of keys' do
-  #     expect(builder.find_node(['query'], root_node)).to eq(root_node[:query])
-  #   end
-
-  #   it 'returns the node from a single key' do
-  #     expect(builder.find_node('query', root_node)).to eq(root_node[:query])
-  #   end
-
-  #   it 'returns the nested node' do
-  #     expect(builder.find_node(['query', 'test'], root_node)).to eq(root_node[:query][:test])
-  #   end
-
-  #   # Fails without hashwithindifferentaccess
-  #   xit 'returns the nested node without hashwithindifferentaccess' do
-  #     expect(builder.find_node(['query', 'test'], root_node2)).to eq(root_node[:query][:test])
-  #   end
-  # end
 
   describe '.blank?' do
     it 'returns true for single args' do
-      expect(builder.blank?(*[])).to eq(true)
-      expect(builder.blank?()).to eq(true)
-      expect(builder.blank?(nil)).to eq(true)
-      expect(builder.blank?(*nil)).to eq(true)
-      expect(builder.blank?([])).to eq(true)
-      expect(builder.blank?({})).to eq(true)
-      expect(builder.blank?('')).to eq(true)
+      expect(utils.blank?(*[])).to eq(true)
+      expect(utils.blank?()).to eq(true)
+      expect(utils.blank?(nil)).to eq(true)
+      expect(utils.blank?(*nil)).to eq(true)
+      expect(utils.blank?([])).to eq(true)
+      expect(utils.blank?({})).to eq(true)
+      expect(utils.blank?('')).to eq(true)
     end
     it 'returns false for single args' do
-      expect(builder.blank?(Object)).to eq(false)
-      expect(builder.blank?('test')).to eq(false)
-      expect(builder.blank?(2)).to eq(false)
+      expect(utils.blank?(Object)).to eq(false)
+      expect(utils.blank?('test')).to eq(false)
+      expect(utils.blank?(2)).to eq(false)
     end
 
     it 'returns true for multiple args' do
-      expect(builder.blank?([[]])).to eq(true)
-      expect(builder.blank?([], [])).to eq(true)
-      expect(builder.blank?({}, [])).to eq(true)
-      expect(builder.blank?(nil, [])).to eq(true)
+      expect(utils.blank?([[]])).to eq(true)
+      expect(utils.blank?([], [])).to eq(true)
+      expect(utils.blank?({}, [])).to eq(true)
+      expect(utils.blank?(nil, [])).to eq(true)
     end
 
     it 'returns false for multiple args' do
-      expect(builder.blank?([], ['s'])).to eq(false)
-      expect(builder.blank?(1,2,3)).to eq(false)
-      expect(builder.blank?('a', 2, 'b')).to eq(false)
+      expect(utils.blank?([], ['s'])).to eq(false)
+      expect(utils.blank?(1,2,3)).to eq(false)
+      expect(utils.blank?('a', 2, 'b')).to eq(false)
     end
   end
 end