+++ CHANGELOG.md
@@ -1,3 +1,32 @@
 # ChangeLog
 
-Due to rapid development of the initial features, the changelog will likely not reflect changes below v0.0.1.
+V 0.0.2
+ - **[Breaking Change]** Interface for bool has changed to use an options hash, previously optional "root_node" arg is now {root: some_root}
+ - **[Breaking Change]** SearchBuilder#node renamed to node?
+ - [Feature] SearchBuilder#size() is available to set the size of the results from ES and is chainable
+ - [Feature] SearchBuilder#size? is available to query the size of the results from ES
+ - [BugFix] Elastic::DSL::Builders::Utils#find_node will return nil/false values if that's what the node maps to, previously it wrongly returned a NodeNotFound Error.
+ - [Feature] More Tests!
+ - [Warning] Queries::Bool - when used with nested `bool`s, it default to merging with the first `bool` added, this can cause problem if you're trying to query like (a || b) && (c || d). Use `options[:append] = true` and add at least one nested `bool` that doesn't have a `should`.
+
+    {query:
+        {bool:
+          {must: [
+            {bool:
+              {should: [
+                condition_a,
+                condition_b,
+              ]},
+            },
+            {bool:
+              {should: [
+                condition_c,
+                condition_d,
+              ]},
+            },
+          ]}
+        }
+      }
+
+V 0.0.1
+ - Due to rapid development of the initial features, the changelog will likely not reflect changes below v0.0.1.